Rename microphone.py to pyaudio.py
